move connection models into a model dir
