+++ src/Harald.WebApi/features/connections/Domain/Model/ChannelId.cs
