+++ src/Harald.WebApi/features/connections/Domain/Model/ChannelType.cs
@@ -1,7 +1,7 @@
 using System;
 using Harald.WebApi.Domain;
 
-namespace Harald.WebApi.features.connections.Domain
+namespace Harald.WebApi.features.connections.Domain.Model
 {
     public class ChannelType : StringSubstitutable
     {
