+++ src/Harald.WebApi/features/connections/Domain/Model/Connection.cs
@@ -1,4 +1,4 @@
-namespace Harald.WebApi.features.connections.Domain
+namespace Harald.WebApi.features.connections.Domain.Model
 {
     public class Connection
     {
