+++ src/Harald.WebApi/features/connections/Infrastructure/driving/api/ConnectionsController.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Threading.Tasks;
 using Harald.WebApi.features.connections.Domain;
+using Harald.WebApi.features.connections.Domain.Model;
 using Harald.WebApi.features.connections.Infrastructure.driving.api.model;
 using Microsoft.AspNetCore.Mvc;
